fix extra_resources.
